+++ app/src/main/java/com/inb/spendly/models/relations/CategoryWithFilteredExpenses.kt
@@ -3,7 +3,7 @@ package com.inb.spendly.models.relations
 data class CategoryWithFilteredExpenses(
     val categoryId: Long,
     val categoryName: String,
-    val categoryColor: Long,
+    val categoryColor: Int,
     val categoryIconId: Int,
     val expenseAmount: Float?
 )
